+++ packages/core/src/components/theme-provider/theme-provider.tsx
@@ -3,7 +3,7 @@
 import type { ReactNode } from 'react';
 import { createContext, memo, useCallback, useContext, useEffect, useMemo, useState } from 'react';
 
-import { RADIUS_FACTOR_VAR_NAME, SCALE_FACTOR_VAR_NAME } from '~/styles/constants';
+import { RADIUS_FACTOR_VAR_NAME, SCALE_FACTOR_VAR_NAME } from '~/styles/global-var.css';
 
 import { THEME_CONFIG, themeInjectScript } from './theme-injector';
 import type { Appearance, Radius, Scaling, ThemeState } from './theme-injector';
