+++ packages/core/src/styles/tokens/size/border-radius.ts
@@ -1,6 +1,6 @@
 import { calc } from '@vanilla-extract/css-utils';
 
-import { radiusFactorVar } from '~/styles/vars.css';
+import { radiusFactorVar } from '~/styles/global-var.css';
 
 export const BORDER_RADIUS = {
     '000': calc.multiply(radiusFactorVar, '0px'),
