+++ packages/core/src/styles/tokens/size/space.ts
@@ -1,6 +1,6 @@
 import { calc } from '@vanilla-extract/css-utils';
 
-import { scaleFactorVar } from '~/styles/vars.css';
+import { scaleFactorVar } from '~/styles/global-var.css';
 
 export const SPACE = {
     '000': calc.multiply(scaleFactorVar, '0px'),
