+++ packages/core/src/styles/tokens/typography/font-size.ts
@@ -1,6 +1,6 @@
 import { calc } from '@vanilla-extract/css-utils';
 
-import { scaleFactorVar } from '~/styles/vars.css';
+import { scaleFactorVar } from '~/styles/global-var.css';
 
 export const FONT_SIZE = {
     '025': calc.multiply(scaleFactorVar, '10px'),
