+++ packages/core/src/styles/tokens/typography/line-height.ts
@@ -1,6 +1,6 @@
 import { calc } from '@vanilla-extract/css-utils';
 
-import { scaleFactorVar } from '~/styles/vars.css';
+import { scaleFactorVar } from '~/styles/global-var.css';
 
 export const LINE_HEIGHT = {
     '025': calc.multiply(scaleFactorVar, '14px'),
