+++ packages/core/src/styles/utils/layer-style.css.ts
@@ -1,7 +1,7 @@
 import type { StyleRule } from '@vanilla-extract/css';
 import { style } from '@vanilla-extract/css';
 
-import type { layerName } from '../constants';
+import type { layerName } from '../layers.css';
 
 export const layerStyle = (
     layer: (typeof layerName)[keyof typeof layerName], // 'vapor-theme' | 'vapor-reset' ...
